ensure commandline argument is emitted by vpr route task
